+++ mos/src/main/java/com/mos/MosApplication.java
@@ -2,8 +2,9 @@ package com.mos;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
 
-@SpringBootApplication
+@SpringBootApplication(exclude = SecurityAutoConfiguration.class)
 public class MosApplication {
 
   public static void main(String[] args) {
