--- mos/src/main/java/com/mos/domain/githuboauth/controller/HomeController.java
@@ -1,14 +1,0 @@
-package com.mos.domain.githuboauth.controller;
-
-import lombok.Getter;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class HomeController {
-
-    @GetMapping("/")
-    public String home() {
-        return "home";
-    }
-}
