--- mos/src/main/java/com/mos/domain/githuboauth/dao/MemberDao.java
@@ -1,14 +1,0 @@
-package com.mos.domain.githuboauth.dao;
-
-import com.mos.vo.Member;
-import com.mos.vo.MemberJoinDto;
-import org.apache.ibatis.annotations.Mapper;
-import org.springframework.stereotype.Component;
-import org.springframework.stereotype.Repository;
-
-@Mapper
-@Component
-public interface MemberDao {
-    public int add(MemberJoinDto joinDto);
-
-}
