--- mos/src/main/java/com/mos/domain/githuboauth/service/MemberService.java
@@ -1,9 +1,0 @@
-package com.mos.domain.githuboauth.service;
-
-import com.mos.vo.Member;
-import com.mos.vo.MemberJoinDto;
-
-public interface MemberService {
-    int add(MemberJoinDto joinDto);
-
-}
