+++ mos/src/main/java/com/mos/domain/member/controller/GithubOAuthController.java
@@ -1,16 +1,11 @@
-package com.mos.domain.githuboauth.controller;
+package com.mos.domain.member.controller;
 
-import com.mos.dao.MemberDao;
-import com.mos.service.GithubOAuthService;
-import com.mos.service.MBMemberService;
-import com.mos.service.MemberService;
-import com.mos.vo.MemberJoinDto;
+import com.mos.domain.member.service.impl.GithubOAuthService;
 import java.io.IOException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 @Controller
@@ -23,6 +18,6 @@ public class GithubOAuthController {
   public String githubLogin(@RequestParam String code, Model model) throws IOException{
     String email = loginService.getAccessToken(code);
     model.addAttribute("email", email);
-    return "member/form";
+    return "auth/form";
   }
 }
