+++ mos/src/main/java/com/mos/domain/member/controller/MemberController.java
@@ -1,13 +1,14 @@
 package com.mos.domain.member.controller;
 
 import com.mos.domain.member.dto.MemberDto;
-import com.mos.domain.member.service.MemberService;
+import com.mos.domain.member.dto.MemberJoinDto;
+import com.mos.domain.member.service.impl.DefaultMemberService;
+import com.mos.domain.member.service.impl.MBMemberService;
 import lombok.RequiredArgsConstructor;
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
-import org.springframework.beans.factory.InitializingBean;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @RequiredArgsConstructor
@@ -15,7 +16,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 @RequestMapping("/member")
 public class MemberController {
 
-  private final MemberService memberService;
+  private final DefaultMemberService memberService;
 
   @GetMapping("findByEmail")
   public void findByEmail(String email) throws Exception{
@@ -26,4 +27,13 @@ public class MemberController {
     memberService.get(member.getEmail());
   }
 
+
+  private final MBMemberService mbMemberService;
+
+
+  @PostMapping("add")
+  public String add(MemberJoinDto joinDto) {
+    mbMemberService.add(joinDto);
+    return "redirect:/";
+  }
 }
