+++ mos/src/main/java/com/mos/domain/member/repository/MemberRepository.java
@@ -1,10 +1,13 @@
 package com.mos.domain.member.repository;
 
 import com.mos.domain.member.dto.MemberDto;
+import com.mos.domain.member.dto.MemberJoinDto;
 import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
 public interface MemberRepository {
+
+  int add(MemberJoinDto joinDto);
 
   MemberDto findByEmail(String email);
 }
