+++ mos/src/main/java/com/mos/domain/member/service/MemberService.java
@@ -5,4 +5,6 @@ import com.mos.domain.member.dto.MemberDto;
 public interface MemberService {
 
   MemberDto get(String email);
+//  int add(MemberJoinDto joinDto);
+
 }
