+++ mos/src/main/resources/application.properties
@@ -40,3 +40,7 @@ kakao.client.id=76cdebf57ec759e3272888d85b54da58
 kakao.client.secret=G21QgA89zA07sO4MPH0LQoMXUwrMz8ba
 kakao.redirect.url=http://localhost:8888/auth/kakao/callback
 
+
+#github OAuth client settings
+spring.security.oauth2.client.registration.github.clientId=Iv1.3954010d06ea3b68
+spring.security.oauth2.client.registration.github.clientSecret=3be9e8332fe59b0fd1cdaf349c57cbec55dbf167
