+++ mos/src/main/resources/templates/auth/form.html
@@ -1,0 +1,39 @@
+<!DOCTYPE html>
+<html lang='en'>
+<head>
+  <meta charset='UTF-8'>
+  <title>비트캠프 데브옵스 5기</title>
+</head>
+<body>
+
+<div data-th-replace="header :: header">머리말</div>
+
+<h1>과제 관리 시스템</h1>
+
+<h2>회원</h2>
+
+<form action="/member/add" data-th-action="@{/member/add}" enctype='multipart/form-data' method='post'>
+  <div>
+    email: <input data-th-value='${email}' name='email' type='text' >
+  </div>
+  <div>
+    이름: <input name='name' type='text'>
+  </div>
+  <div>
+    belong: <input name='belong' type='text'>
+  </div>
+  <div>
+    career: <input name='career' type='text'>
+  </div>
+  <div>
+    jobGroup: <input name='jobGroup' type='text'>
+  </div>
+  <div>
+    <button>등록</button>
+  </div>
+</form>
+
+<div data-th-replace="footer :: footer">꼬리말</div>
+
+</body>
+</html>
