+++ mos/src/main/resources/templates/auth/login.html
@@ -10,6 +10,7 @@
 <div data-th-replace="header::header">머리말</div>
 
 <div class="choose_login">
+  <a href="https://github.com/login/oauth/authorize?client_id=Iv1.3954010d06ea3b68&scope=user:email"/>
   <img src="/img/logo_text.png">
   <div class="container">
     <div class="login-option">
